requirements.txt
astroid==2.4.2
CacheControl==0.12.6
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.15
django-cors-headers==3.4.0
django-reset-migrations==0.4.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
et-xmlfile==1.0.1
firebase-admin==4.4.0
google-api-core==1.22.2
google-api-python-client==1.12.2
google-auth==1.21.3
google-auth-httplib2==0.0.4
google-cloud-core==1.4.1
google-cloud-firestore==1.9.0
google-cloud-storage==1.31.2
google-crc32c==1.0.0
google-resumable-media==1.0.0
googleapis-common-protos==1.52.0
grpcio==1.32.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
isort==5.5.3
jdcal==1.4.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
msgpack==1.0.0
openpyxl==3.0.5
Pillow==7.2.0
protobuf==3.13.0
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.10
whitenoise==5.2.0
wrapt==1.12.1
